+++ Models/Candidtate.cs
@@ -4,10 +4,10 @@ namespace FindJobWebApi.Models
 {
     public class Candidtate
     {
-        public ulong Id { get; set; }
-        public ulong VacancyId { get; set; }
+        public int Id { get; set; }
+        public int VacancyId { get; set; }
         public virtual Vacancy Vacancy { get; set; }
-        public ulong UserId { get; set; }
+        public int UserId { get; set; }
         public virtual User User { get; set; }
     }
 }
